* 'master' of https://bitbucket.org/lei-isep/lapr3-2020-g033:
  AUCMN1-212 #changes to database
  AUCMN1-52 MD Diagrams Update
  AUCMN1-211 AUCMN1-212 #changes to database + added new functions
  AUCMN1-244 AUCMN1-240 #comment Worked on UC7 and UC9
  AUCMN1-258 #progress [development]
  resolve merge conflits
  problems resolved
  no message
  AUCMN1-220 #progress continuing the tests
  AUCMN1-213 #comment solved code smells

+++ src/main/java/lapr/project/controller/PharmacyController.java
@@ -50,7 +50,7 @@ public class PharmacyController {
         return pharmacyDataHandler.getPharmacyByID(id);
     }
 
-    public boolean addPark(int maxCapacity, int maxChargingPlaces, int power, int pharmacyID, int idParktype) {
+    public boolean addPark(int maxCapacity, int maxChargingPlaces, double power, int pharmacyID, int idParktype) {
         try {
             if(getPark(pharmacyID,idParktype)!=null){
                 return false;
@@ -71,7 +71,7 @@ public class PharmacyController {
 
 
 
-    public boolean registerPharmacyandPark(String name, double latitude, double longitude, String street, int doorNumber, String zipCode, String locality, int maxCpacity, int maxChargingCapacity, int power,int idParkType, String emailAdmin, String emailP) {
+    public boolean registerPharmacyandPark(String name, double latitude, double longitude, String street, int doorNumber, String zipCode, String locality, int maxCpacity, int maxChargingCapacity, double power,int idParkType, String emailAdmin, String emailP) {
            try{
                Address add = new Address(latitude, longitude, street, doorNumber, zipCode, locality);
                boolean addCheck = addressDataHandler.addAddress(add);
